Moves view files into the view folder
