+++ package.json
@@ -39,8 +39,8 @@
       {
         "command": "sampleQuery.Click",
         "title": "click"
-	  },
-	  {
+      },
+      {
         "command": "history.populate",
         "title": "populate"
       }
