+++ src/view/sampleQueryProvider.ts
